backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
python-jose[cryptography]
google-auth
google-auth-httplib2
langchain
langchain-google-genai
google-genai
apscheduler
google-cloud-pubsub
firebase-admin
python-multipart
aiofiles
httpx
passlib[bcrypt]
